Refactored world.py to msg_receive.py.
