+++ commands/reload.py
@@ -1,5 +1,5 @@
-import world
+import msg_receive
 
 def run(connection, privmsg):
-    connection.valid_commands = world.loadCommands(connection.commandpath)
+    connection.valid_commands = msg_receive.loadCommands(connection.commandpath)
     return
